requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.1
backoff==2.2.1
beautifulsoup4==4.12.2
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.4
coloredlogs==15.0.1
confection==0.0.4
cryptography==41.0.1
cymem==2.0.7
dataclasses-json==0.5.8
decorator==5.1.1
dill==0.3.6
dnspython==2.3.0
docopt==0.6.2
duckdb==0.8.1
elastic-transport==8.4.0
elasticsearch==8.8.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
faiss-cpu==1.7.4
fastapi==0.97.0
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
geojson==2.5.0
google-api-core==2.11.1
google-api-python-client==2.70.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-tools==1.56.0
h11==0.14.0
h2==4.1.0
hnswlib==0.7.0
hpack==4.0.0
httpcore==0.17.2
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.15.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
jaraco.context==4.3.0
Jinja2==3.1.2
joblib==1.3.0
langchain==0.0.231
langchainplus-sdk==0.0.20
langcodes==3.3.0
loguru==0.7.0
lz4==4.3.2
manifest-ml==0.0.1
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.1
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pathy==0.10.2
pinecone-client==2.2.2
pipreqs==0.4.13
posthog==3.0.1
preshed==3.0.8
protobuf==4.23.3
pulsar-client==3.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pyowm==3.3.0
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
qdrant-client==0.11.10
redis==4.6.0
regex==2023.6.3
requests==2.31.0
rsa==4.9
safetensors==0.3.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.48
sqlitedict==2.1.0
srsly==2.4.6
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
thinc==8.1.10
tiktoken==0.4.0
tokenizers==0.13.3
torch==1.13.1
tqdm==4.65.0
transformers==4.30.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
wasabi==1.1.2
watchfiles==0.19.0
weaviate-client==3.21.0
websockets==11.0.3
wikipedia==1.4.0
wolframalpha==5.0.0
xmltodict==0.13.0
yarg==0.1.9
yarl==1.9.2
zstandard==0.21.0
